pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "savile"
version = "0.1.0"
description = "System for Agentic Versioning, Intelligence, and Logical Evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "mcp>=0.1.0",
    "pyyaml>=6.0",
    "GitPython>=3.1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0"
]

[project.scripts]
savile = "savile.cli:app"
